Add .gitignore, MANIFEST.SKIP, then update MANIFEST
While working on https://github.com/jo-37/List-Gen/pull/1, I had
occasion several times to say 'git status', first having called either
'perl Makefile.PL' or 'perl ./Build'.  That was listing several
files not stored in the repository, but which don't need to be there --
though they have to be accounted for in the tarball -- and which should
be ignored by git.

Conversely, a '.gitignore' should not go in the tarball and hence should
be in a MANIFEST.SKIP file.  I copied in a .gitignore and a
MANIFEST.SKIP from one of my CPAN distros.